pyproject.toml
[project]
name = "contract-automation"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "Neuronspeeed", email = "neuronspeeed@gmail.com" }
]
dependencies = [
    "langgraph>=0.2.34",
    "typing>=3.7.4.3",
    "ipynb>=0.5.1",
    "ipykernel>=6.29.5",
    "python-dotenv>=1.0.1",
    "langchain>=0.3.1",
    "openai>=1.51.0",
    "langchain-openai>=0.2.1",
    "pydantic>=2.9.2",
    "pytesseract>=0.3.13",
    "pillow>=10.4.0",
    "config>=0.5.1",
    "aiofiles>=24.1.0",
    "langchain-community>=0.3.3",
    "pdf2image>=1.17.0",
    "pymupdf>=1.24.11",
    "pypdf2>=3.0.1",
    "easyocr>=1.7.2",
    "instructor>=1.5.2",
    "pdfplumber>=0.11.4",
    "pytest>=8.3.3",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    "ratelimit>=2.2.1",
    "pytest-tornasync>=0.6.0.post2",
    "pytest-trio>=0.8.0",
    "pytest-twisted>=1.14.3",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/contract_automation"]

[tool.pytest.ini_options]
addopts = "-v"
testpaths = ["tests"]
asyncio_mode = "auto"
markers = [
    "asyncio: mark test as requiring asyncio",
]
